+++ HW5/src/kernel4.cu
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-__global__ void mandelKernel(int *device_data, int lowerX, int lowerY, int stepX, int stepY, int resX, int maxIterations){
+__global__ void mandelKernel(int *device_data, float lowerX, float lowerY, float stepX, float stepY, int resX, int maxIterations){
     // To avoid error caused by the floating number, use the following pseudo code
     //
     // float x = lowerX + thisX * stepX;
